+++ src/main/java/com/gilmanov/sbootpp312/controllers/UserController.java
@@ -1,7 +1,7 @@
 package com.gilmanov.sbootpp312.controllers;
 
 import com.gilmanov.sbootpp312.model.User;
-import com.gilmanov.sbootpp312.services.UserService;
+import com.gilmanov.sbootpp312.services.UserServiceImpl;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
@@ -10,15 +10,15 @@ import org.springframework.web.bind.annotation.*;
 @Controller
 public class UserController {
 
-    private final UserService userService;
+    private final UserServiceImpl userService;
 
-    public UserController(UserService userService) {
+    public UserController(UserServiceImpl userService) {
         this.userService = userService;
     }
 
     @GetMapping("/users")
     public String allUsers(Model model) {
-        model.addAttribute("users", userService.getAllUsers());
+        model.addAttribute("users2", userService.getAllUsers());
         return "users";
     }
