+++ src/main/resources/templates/users.html
@@ -1,5 +1,4 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://thymeleaf.org" lang="en">
+<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org" lang="en">
 <html lang="en">
 <head>
     <meta charset="UTF-8">
@@ -18,7 +17,7 @@
         </tr>
         </thead>
         <tbody>
-        <tr th:each="user :${users}">
+        <tr th:each="user :${users2}">
             <td th:text="${user.getName()}"/>
             <td th:text="${user.getSurname()}"/>
             <td th:text="${user.getAge()}"/>
